Extract colors and typography to files
