+++ class/resources/colors.ts
@@ -1,0 +1,28 @@
+export type ColorCode = `\x1b[${string}m`;
+
+export const colorCodes = {
+  black: "\x1b[30m",
+  red: "\x1b[31m",
+  green: "\x1b[32m",
+  yellow: "\x1b[33m",
+  blue: "\x1b[34m",
+  magenta: "\x1b[35m",
+  cyan: "\x1b[36m",
+  white: "\x1b[37m",
+} satisfies Record<string, ColorCode>;
+
+export const bgColorCodes = {
+  //background color
+  black: "\x1b[40m",
+  red: "\x1b[41m",
+  green: "\x1b[42m",
+  yellow: "\x1b[43m",
+  blue: "\x1b[44m",
+  magenta: "\x1b[45m",
+  cyan: "\x1b[46m",
+  white: "\x1b[47m",
+} satisfies Record<Color, ColorCode>;
+
+export const resetColor: ColorCode = "\x1b[0m";
+
+export type Color = keyof typeof colorCodes;
